scripts done

+++ scripts/aws_ecs_benchmarks_cdn.sh
@@ -19,9 +19,10 @@ ecs deploy --region us-east-2 hotshot hotshot_centralized -i centralized ghcr.io
 ecs deploy --region us-east-2 hotshot hotshot_centralized -c centralized ${orchestrator_url} # http://172.31.8.82:4444
 
 # runstart keydb
-docker run --rm -p 0.0.0.0:6379:6379 eqalpha/keydb
+# docker run --rm -p 0.0.0.0:6379:6379 eqalpha/keydb
 # server1: broker and marshal
 just async_std example cdn-marshal -- -d redis://localhost:6379 -b 9000 &
+# remember to sleep enough time if it's built first time
 just async_std example cdn-broker -- -d redis://localhost:6379 \
     --public-bind-endpoint 0.0.0.0:1740 \
     --public-advertise-endpoint local_ip:1740 \
@@ -33,37 +34,45 @@ just async_std example cdn-broker -- -d redis://localhost:6379 \
 # total_nodes, da_committee_size, transactions_per_round, transaction_size = 100, 10, 1, 4096
 # for iteration of assignment
 # see `aws_ecs_nginx_benchmarks.sh` for an example
-for total_nodes in 10 50 100
+for total_nodes in 10 # 50 100
 do
-    for da_committee_size in 10 50 100
+    for da_committee_size in 5 #10 50 100
     do
         if [ $da_committee_size -le $total_nodes ]
         then
-            for transactions_per_round in 1 10 50 100
+            for transactions_per_round in 1 # 10 50 100
             do
-                for transaction_size in 512 4096 # see large transaction size in aws_ecs_nginx_benchmarks.sh
+                for transaction_size in 512 4096 1000000
                 do
-                    rounds=100
+                    for fixed_leader_for_gpuvid in 1 #5 10
+                    do
+                        if [ $fixed_leader_for_gpuvid -le $da_committee_size ]
+                        then
+                            rounds=100
 
-                    # start orchestrator
-                    just async_std example orchestrator-push-cdn -- --config_file ./crates/orchestrator/run-config.toml \
-                                                                    --orchestrator_url http://0.0.0.0:4444 \
-                                                                    --total_nodes ${total_nodes} \
-                                                                    --da_committee_size ${da_committee_size} \
-                                                                    --transactions_per_round ${transactions_per_round} \
-                                                                    --transaction_size ${transaction_size} \
-                                                                    --rounds ${rounds} \
-                                                                    --commit_sha test &
-                    sleep 30
+                            # start orchestrator
+                            just async_std example orchestrator-push-cdn -- --config_file ./crates/orchestrator/run-config.toml \
+                                                                            --orchestrator_url http://0.0.0.0:4444 \
+                                                                            --total_nodes ${total_nodes} \
+                                                                            --da_committee_size ${da_committee_size} \
+                                                                            --transactions_per_round ${transactions_per_round} \
+                                                                            --transaction_size ${transaction_size} \
+                                                                            --rounds ${rounds} \
+                                                                            --fixed_leader_for_gpuvid ${fixed_leader_for_gpuvid} \
+                                                                            --commit_sha cdn_no_gpu &
+                            sleep 30
 
-                    # start validators
-                    ecs scale --region us-east-2 hotshot hotshot_centralized ${total_nodes} --timeout -1
-                    sleep $(($rounds + $total_nodes))
+                            # start validators
+                            ecs scale --region us-east-2 hotshot hotshot_centralized ${total_nodes} --timeout -1
+                            sleep $(( ($rounds + $total_nodes) * $transactions_per_round ))
 
-                    # kill them
-                    ecs scale --region us-east-2 hotshot hotshot_centralized 0 --timeout -1
-                    sleep 1m
-                    for pid in $(ps -ef | grep "orchestrator" | awk '{print $2}'); do kill -9 $pid; done
+                            # kill them
+                            ecs scale --region us-east-2 hotshot hotshot_centralized 0 --timeout -1
+                            sleep 1m
+                            for pid in $(ps -ef | grep "orchestrator" | awk '{print $2}'); do kill -9 $pid; done
+                            sleep 10
+                        fi
+                    done
                 done
             done
         fi
@@ -71,3 +80,5 @@ do
 done
 
 # shut down all related threads
+for pid in $(ps -ef | grep "cdn-broker" | awk '{print $2}'); do kill -9 $pid; done
+for pid in $(ps -ef | grep "cdn-marshal" | awk '{print $2}'); do kill -9 $pid; done
